+++ app/uk/gov/hmrc/mobiletaxcreditssummary/controllers/TaxCreditsSummaryController.scala
@@ -18,11 +18,9 @@ package uk.gov.hmrc.mobiletaxcreditssummary.controllers
 
 import javax.inject.{Inject, Named, Singleton}
 import play.api._
-import play.api.libs.json.Json
 import play.api.libs.json.Json.{obj, toJson}
 import play.api.mvc._
 import uk.gov.hmrc.api.controllers._
-import uk.gov.hmrc.api.sandbox.FileResource
 import uk.gov.hmrc.api.service.Auditor
 import uk.gov.hmrc.auth.core.AuthConnector
 import uk.gov.hmrc.domain.Nino
@@ -32,10 +30,10 @@ import uk.gov.hmrc.mobiletaxcreditssummary.domain._
 import uk.gov.hmrc.mobiletaxcreditssummary.domain.userdata._
 import uk.gov.hmrc.mobiletaxcreditssummary.services.LiveTaxCreditsSummaryService
 import uk.gov.hmrc.play.HeaderCarrierConverter.fromHeadersAndSession
+import uk.gov.hmrc.play.audit.AuditExtensions.auditHeaderCarrier
 import uk.gov.hmrc.play.audit.http.connector.AuditConnector
 import uk.gov.hmrc.play.audit.model.ExtendedDataEvent
 import uk.gov.hmrc.play.bootstrap.controller.BaseController
-import uk.gov.hmrc.play.audit.AuditExtensions.auditHeaderCarrier
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
@@ -75,41 +73,6 @@ trait TaxCreditsSummaryController extends BaseController {
 
   def taxCreditsSummary(nino: Nino, journeyId: Option[String] = None): Action[AnyContent]
 
-}
-
-@Singleton
-class SandboxTaxCreditsSummaryController() extends TaxCreditsSummaryController with FileResource with HeaderValidator {
-  val shuttering =
-    ConfiguredShuttering(
-      shuttered = true,
-      "Service Unavailable",
-      Seq(
-        "You'll be able to use the app to manage your tax credits at 9am on Monday 29 May 2017.",
-        "Go to GOV UK to <a href=“https://www.gov.uk“>manage your tax credits online</a>."))
-
-  override final def taxCreditsSummary(nino: Nino, journeyId: Option[String] = None): Action[AnyContent] =
-    validateAccept(acceptHeaderValidationRules).async {
-      implicit request =>
-        Future successful (request.headers.get("SANDBOX-CONTROL") match {
-          case Some("NON-TAX-CREDITS-USER") => Ok(toJson(TaxCreditsSummaryResponse(taxCreditsSummary = None)))
-          case Some("EXCLUDED-TAX-CREDITS-USER") => Ok(toJson(TaxCreditsSummaryResponse(excluded = true, taxCreditsSummary = None)))
-          case Some("ERROR-401") => Unauthorized
-          case Some("ERROR-403") => Forbidden
-          case Some("ERROR-500") => InternalServerError
-          case Some("SHUTTERED") => ServiceUnavailable(toJson(shuttering))
-          case Some("CLAIMANTS_FAILURE") =>
-            val resource: String = findResource(s"/resources/taxcreditssummary/${nino.value}.json")
-              .getOrElse(throw new IllegalArgumentException("Resource not found!"))
-            val taxCreditsSummary: TaxCreditsSummary = TaxCreditsSummary(Json.parse(resource).as[TaxCreditsSummary].paymentSummary, None)
-            val response = TaxCreditsSummaryResponse(excluded = false, Some(taxCreditsSummary))
-            Ok(toJson(response))
-          case _ => //TAX-CREDITS-USER
-            val resource: String = findResource(s"/resources/taxcreditssummary/${nino.value}.json")
-              .getOrElse(throw new IllegalArgumentException("Resource not found!"))
-            val response = TaxCreditsSummaryResponse(excluded = false, Some(Json.parse(resource).as[TaxCreditsSummary]))
-            Ok(toJson(response))
-        })
-    }
 }
 
 @Singleton
