+++ app/uk/gov/hmrc/mobiletaxcreditssummary/controllers/action/Authorisation.scala
@@ -16,22 +16,28 @@
 
 package uk.gov.hmrc.mobiletaxcreditssummary.controllers.action
 
+import play.api.Logger
+import play.api.libs.json.Json
+import play.api.mvc.{ActionBuilder, Request, Result, Results}
+import uk.gov.hmrc.api.controllers._
 import uk.gov.hmrc.auth.core.retrieve.Retrievals._
 import uk.gov.hmrc.auth.core.retrieve.~
 import uk.gov.hmrc.auth.core.{AuthorisedFunctions, Enrolment, EnrolmentIdentifier}
 import uk.gov.hmrc.domain.Nino
 import uk.gov.hmrc.http.HeaderCarrier
-import uk.gov.hmrc.mobiletaxcreditssummary.controllers.{AccountWithLowCL, FailToMatchTaxIdOnAuth, NinoNotFoundOnAccount}
+import uk.gov.hmrc.mobiletaxcreditssummary.controllers.{AccountWithLowCL, FailToMatchTaxIdOnAuth, NinoNotFoundOnAccount, _}
+import uk.gov.hmrc.play.HeaderCarrierConverter
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
 
 case class Authority(nino: Nino)
 
-trait Authorisation extends AuthorisedFunctions {
+trait Authorisation extends Results with AuthorisedFunctions {
 
   val confLevel: Int
 
+  lazy val requiresAuth: Boolean = true
   lazy val ninoNotFoundOnAccount = new NinoNotFoundOnAccount
   lazy val failedToMatchNino = new FailToMatchTaxIdOnAuth
   lazy val lowConfidenceLevel = new AccountWithLowCL
@@ -47,5 +53,43 @@ trait Authorisation extends AuthorisedFunctions {
         case None ~ _ =>
           throw ninoNotFoundOnAccount
       }
+  }
+
+  def invokeAuthBlock[A](request: Request[A], block: Request[A] => Future[Result], taxId: Option[Nino]): Future[Result] = {
+    implicit val hc: HeaderCarrier = HeaderCarrierConverter.fromHeadersAndSession(request.headers, None)
+
+    grantAccess(taxId.getOrElse(Nino(""))).flatMap { _ =>
+      block(request)
+    }.recover {
+      case _: uk.gov.hmrc.http.Upstream4xxResponse =>
+        Logger.info("Unauthorized! Failed to grant access since 4xx response!")
+        Unauthorized(Json.toJson(ErrorUnauthorizedMicroService))
+
+      case _: NinoNotFoundOnAccount =>
+        Logger.info("Unauthorized! NINO not found on account!")
+        Unauthorized(Json.toJson(ErrorUnauthorizedNoNino))
+
+      case _: FailToMatchTaxIdOnAuth =>
+        Logger.info("Forbidden! Failure to match URL NINO against Auth NINO")
+        Forbidden(Json.toJson(ErrorForbidden))
+
+      case _: AccountWithLowCL =>
+        Logger.info("Unauthorized! Account with low CL!")
+        Unauthorized(Json.toJson(ErrorUnauthorizedLowCL))
+    }
+  }
+}
+
+trait AccessControl extends HeaderValidator with Authorisation {
+
+  def validateAcceptWithAuth(rules: Option[String] => Boolean, taxId: Option[Nino]): ActionBuilder[Request] = new ActionBuilder[Request] {
+
+    def invokeBlock[A](request: Request[A], block: Request[A] => Future[Result]): Future[Result] = {
+      if (rules(request.headers.get("Accept"))) {
+        if (requiresAuth) invokeAuthBlock(request, block, taxId)
+        else block(request)
+      }
+      else Future.successful(Status(ErrorAcceptHeaderInvalid.httpStatusCode)(Json.toJson(ErrorAcceptHeaderInvalid)))
+    }
   }
 }
